+++ Middlewares/ST/netxduo/addons/azure_iot/azure_iot_security_module/iot-security-module-core/src/serializer/serializer_private.h
@@ -1,16 +1,17 @@
-/**************************************************************************/
-/*                                                                        */
-/*       Copyright (c) Microsoft Corporation. All rights reserved.        */
-/*                                                                        */
-/*       This software is licensed under the Microsoft Software License   */
-/*       Terms for Microsoft Azure RTOS. Full text of the license can be  */
-/*       found in the LICENSE file at https://aka.ms/AzureRTOS_EULA       */
-/*       and in the root directory of this software.                      */
-/*                                                                        */
-/**************************************************************************/
+/*******************************************************************************/
+/*                                                                             */
+/* Copyright (c) Microsoft Corporation. All rights reserved.                   */
+/*                                                                             */
+/* This software is licensed under the Microsoft Software License              */
+/* Terms for Microsoft Azure Defender for IoT. Full text of the license can be */
+/* found in the LICENSE file at https://aka.ms/AzureDefenderForIoT_EULA        */
+/* and in the root directory of this software.                                 */
+/*                                                                             */
+/*******************************************************************************/
 
 #ifndef SERIALIZER_PRIVATE_H
 #define SERIALIZER_PRIVATE_H
+#include <asc_config.h>
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -21,6 +22,7 @@
 #include "asc_security_core/serializer.h"
 
 struct serializer {
+    /* This macro must be first in object */
     COLLECTION_INTERFACE(serializer_t);
     
     flatcc_builder_t builder;
@@ -40,6 +42,6 @@ struct serializer {
  * @return  ASC_RESULT_OK on success,
  *          ASC_RESULT_EXCEPTION otherwise
  */
-asc_result_t serializer_event_start(serializer_t *serializer, uint32_t timestamp, uint32_t collection_interval);
+asc_result_t serializer_event_start(serializer_t *serializer, unsigned long timestamp, unsigned long collection_interval);
 
 #endif /* SERIALIZER_PRIVATE_H */
